requirements.txt
python-dotenv==0.21.0
selenium==4.6.0
webdriver_manager==3.8.4
